renamed files and created folder structure in View to organize files
