+++ open-fin-al/src/View/App/Loaded.jsx
@@ -18,11 +18,11 @@ import Home from "../Home";
 import Portfolio from "../Portfolio";
 import { Analysis } from "../Analysis";
 import BuyReport from "../BuyReport";
-import { TimeSeries } from "../TimeSeriesPage";
-import { News } from "../NewsPage";
+import { TimeSeries } from "../Stock";
+import { News } from "../News";
 import { Learn } from "../Learn";
-import { LearningModuleDetails } from "../LearningModuleDetails";
-import { LearningModulePage } from "../LearningModulePage";
+import { LearningModuleDetails } from "../LearningModule/LearningModuleDetails";
+import { LearningModulePage } from "../LearningModule/LearningModulePage";
 import logo from "../../Asset/Image/logo.png";
 import logoNoText from "../../Asset/Image/openfinal_logo_no_text.png";
 import logoNoTextDark from "../../Asset/Image/openfinal_logo_dark_no_text.png";
