+++ open-fin-al/src/View/App/Preparing.jsx
