+++ open-fin-al/src/View/LearningModule/LearningModuleDetails.jsx
