+++ open-fin-al/src/View/LearningModule/Slideshow/Stocks/Slide2.html
