+++ open-fin-al/src/View/News/NewsListing.jsx
