+++ open-fin-al/src/View/Stock/TimeSeriesChart.jsx
